pyproject.toml
[tool.poetry]
name = "bank-app"
version = "0.1.0"
description = ""
authors = ["Mikhail Lemdyanov <lacoustique@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'


[tool.isort]
line_length = 119


[tool.black]
line-length = 119
exclude = '.git'
